Fix CI deployment false negative (#2)
PowerShell was failing the deployment script because `git push` as the
last command was writing to the standard error stream even though it
was terminating successfull with an exit code of zero. Redirecting via
`2>&1` didn't help either so this fix uses the `--quiet` switch to
suppress all output and runs a final `git show` to dump the last
commit.